+++ src/zope/app/file/browser/image.py
@@ -13,7 +13,6 @@
 ##############################################################################
 """Define view component for naive file editing.
 
-$Id$
 """
 __docformat__ = 'restructuredtext'