+++ src/zope/app/file/image.py
@@ -13,14 +13,13 @@
 ##############################################################################
 """Image content type implementation
 
-$Id$
 """
 __docformat__ = 'restructuredtext'
 
 import struct
-from cStringIO import StringIO
+from io import BytesIO
 
-from zope.interface import implements
+from zope.interface import implementer
 from zope.size.interfaces import ISized
 from zope.size import byteDisplay
 from zope.contenttype import guess_content_type
@@ -29,13 +28,14 @@ from zope.app.file.i18n import ZopeMessageFactory as _
 from zope.app.file.file import File
 from zope.app.file.interfaces import IImage
 
-class Image(File):
-    implements(IImage)
 
-    def __init__(self, data=''):
+@implementer(IImage)
+class Image(File):
+
+    def __init__(self, data=b''):
         '''See interface `IFile`'''
-        self.contentType, self._width, self._height = getImageInfo(data)
-        self.data = data
+        contentType, self._width, self._height = getImageInfo(data)
+        super(Image, self).__init__(data, contentType)
 
     def _setData(self, data):
         super(Image, self)._setData(data)
@@ -50,8 +50,9 @@ class Image(File):
 
     data = property(File._getData, _setData)
 
+
+@implementer(ISized)
 class ImageSized(object):
-    implements(ISized)
 
     def __init__(self, image):
         self._image = image
@@ -77,6 +78,7 @@ class ImageSized(object):
         # i18nextract and never show up in message catalogs
         return _(byte_size + ' ${width}x${height}', mapping=mapping)
 
+
 class FileFactory(object):
 
     def __init__(self, context):
@@ -84,9 +86,9 @@ class FileFactory(object):
 
     def __call__(self, name, content_type, data):
         if not content_type and data:
-            content_type, width, height = getImageInfo(data)
+            content_type, _width, _height = getImageInfo(data)
         if not content_type:
-            content_type, encoding = guess_content_type(name, data, '')
+            content_type, _encoding = guess_content_type(name, data, '')
 
         if content_type.startswith('image/'):
             return Image(data)
@@ -94,32 +96,32 @@ class FileFactory(object):
         return File(data, content_type)
 
 def getImageInfo(data):
-    data = str(data)
+    data = bytes(data)
     size = len(data)
     height = -1
     width = -1
     content_type = ''
 
     # handle GIFs
-    if (size >= 10) and data[:6] in ('GIF87a', 'GIF89a'):
+    if (size >= 10) and data[:6] in (b'GIF87a', b'GIF89a'):
         # Check to see if content_type is correct
         content_type = 'image/gif'
-        w, h = struct.unpack("<HH", data[6:10])
+        w, h = struct.unpack(b"<HH", data[6:10])
         width = int(w)
         height = int(h)
 
     # See PNG 2. Edition spec (http://www.w3.org/TR/PNG/)
     # Bytes 0-7 are below, 4-byte chunk length, then 'IHDR'
     # and finally the 4-byte width, height
-    elif ((size >= 24) and data.startswith('\211PNG\r\n\032\n')
-          and (data[12:16] == 'IHDR')):
+    elif ((size >= 24) and data.startswith(b'\211PNG\r\n\032\n')
+          and (data[12:16] == b'IHDR')):
         content_type = 'image/png'
-        w, h = struct.unpack(">LL", data[16:24])
+        w, h = struct.unpack(b">LL", data[16:24])
         width = int(w)
         height = int(h)
 
     # Maybe this is for an older PNG version.
-    elif (size >= 16) and data.startswith('\211PNG\r\n\032\n'):
+    elif (size >= 16) and data.startswith(b'\211PNG\r\n\032\n'):
         # Check to see if we have the right content type
         content_type = 'image/png'
         w, h = struct.unpack(">LL", data[8:16])
@@ -127,9 +129,9 @@ def getImageInfo(data):
         height = int(h)
 
     # handle JPEGs
-    elif (size >= 2) and data.startswith('\377\330'):
+    elif (size >= 2) and data.startswith(b'\377\330'):
         content_type = 'image/jpeg'
-        jpeg = StringIO(data)
+        jpeg = BytesIO(data)
         jpeg.read(2)
         b = jpeg.read(1)
         try:
@@ -140,10 +142,10 @@ def getImageInfo(data):
                 while (ord(b) == 0xFF): b = jpeg.read(1)
                 if (ord(b) >= 0xC0 and ord(b) <= 0xC3):
                     jpeg.read(3)
-                    h, w = struct.unpack(">HH", jpeg.read(4))
+                    h, w = struct.unpack(b">HH", jpeg.read(4))
                     break
                 else:
-                    jpeg.read(int(struct.unpack(">H", jpeg.read(2))[0])-2)
+                    jpeg.read(int(struct.unpack(b">H", jpeg.read(2))[0])-2)
                 b = jpeg.read(1)
             width = int(w)
             height = int(h)
@@ -153,10 +155,10 @@ def getImageInfo(data):
             pass
 
     # handle BMPs
-    elif (size >= 30) and data.startswith('BM'):
-        kind = struct.unpack("<H", data[14:16])[0]
+    elif (size >= 30) and data.startswith(b'BM'):
+        kind = struct.unpack(b"<H", data[14:16])[0]
         if kind == 40: # Windows 3.x bitmap
             content_type = 'image/x-ms-bmp'
-            width, height = struct.unpack("<LL", data[18:26])
+            width, height = struct.unpack(b"<LL", data[18:26])
 
     return content_type, width, height